# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Syrian Car Market API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vehicle-reports/lookup:
    post:
      operationId: lookupVehicle
      tags: [vehicle-reports]
      summary: Look up vehicle information by VIN, plate, or chassis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleLookupBody"
      responses:
        "200":
          description: Vehicle report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleReport"
        "404":
          description: Vehicle not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    VehicleLookupBody:
      type: object
      properties:
        vin:
          type: string
        plateNumber:
          type: string
        chassisNumber:
          type: string
    VehicleReport:
      type: object
      properties:
        id:
          type: integer
        vin:
          type: ["string", "null"]
        plateNumber:
          type: ["string", "null"]
        chassisNumber:
          type: ["string", "null"]
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
        countryOfOrigin:
          type: ["string", "null"]
        engineSize:
          type: ["string", "null"]
        fuelType:
          type: string
        transmission:
          type: string
        engineCapacity:
          type: ["string", "null"]
        horsepower:
          type: ["integer", "null"]
        fuelConsumption:
          type: ["string", "null"]
        driveType:
          type: ["string", "null"]
        accidentCount:
          type: integer
        hasMajorRepairs:
          type: boolean
        hasStructuralDamage:
          type: boolean
        airbagDeployed:
          type: boolean
        mileageHistory:
          type: array
          items:
            $ref: "#/components/schemas/MileageRecord"
        ownershipCount:
          type: integer
        registrationRegion:
          type: ["string", "null"]
        damageStatus:
          type: string
          enum: [clean, minor, serious]
        aiSummary:
          type: ["string", "null"]
      required:
        - id
        - brand
        - model
        - year
        - fuelType
        - transmission
        - accidentCount
        - hasMajorRepairs
        - hasStructuralDamage
        - airbagDeployed
        - mileageHistory
        - ownershipCount
        - damageStatus
    MileageRecord:
      type: object
      properties:
        year:
          type: integer
        mileage:
          type: integer
      required:
        - year
        - mileage
...